Refactor: Join string and string_parser.
